requirements.txt
jupyter==1.0.0
limit-order-book==2.0.0
python-bitcoinlib==0.11.0
